moved full app functionality tests to test_converter
